+++ CHANGELOG.md
@@ -239,8 +239,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
   - From\<&[u8]\> for Der
   - From\<&ContentsRef\> for Contents
   - From\<&[u8; N]\> for Contents
+  - From\<&str\> for Content
+  - From\<Vec\<u8\>\> for Content
+  - From\<String\> for Content
   - From\<&[u8; N]\> for &ContentsRef
   - From\<&mut [u8; N]\> for &mut ContentsRef
+  - From\<&str\> for &ContentsRef
 - Create the following functions and methods
   - Ber::from\_vec\_unchecked()
   - BerRef::eoc()